make Makie a weakdep
