+++ ext/AURORA_viz.jl
@@ -1,0 +1,8 @@
+module AURORA_viz
+
+using AURORA
+include("src/animations/plot.jl")
+include("src/animations/animate.jl")
+# export animate_IeztE_3Dzoft
+
+end
